LF-4154 move folder to Form
